mark: be careful not to free a mark twice.
If a pane that owns a mark has the document under
it freed, all the marks will get freed, but the pane
won't get told.
I guess it could set up a notificaiton, but for now
just allow it to 'free' the mark again.

This is relevant to 'q' in a directory listing.
doc-rendering holds a mark temporarily, but by
the time it wants to free it, it is already gone.

Signed-off-by: NeilBrown <neil@brown.name>