c-mode: include "Enter" in normal undo processing.
I want the same rules that gather multiple inserts into
a single undo, to apply to Enter when typed in C-mode,
which adds an indent.

So create a new Interactive:insert (and matching Interactive:delete)
which emacs supports and which does the right thing, and have
handle:Enter call it.

Might want something similar for backspace.  Probably not
for TAB.

Signed-off-by: NeilBrown <neil@brown.name>