Add mark_make_first() and mark_make_last()
There can be several mark at the same location in a document.
Sometimes it matter what order they are in.
When it does, mark_make_first() can be used to put a mark
first in order, and mark_make_last() can make it last.

When finer control is needed, mark_to_mark() can be used, it always
changes the order of (different) marks.

Signed-off-by: NeilBrown <neil@brown.name>