server: handle document being killed.
If the server opens a document in a -t window, and
the document is killed, the the window cannot be closed - as it refuses
to close until the document is done, and it cannot be "done" any
more.

So get notifications when the document is closed, and treat
that just like the document being marked 'done'.

Signed-off-by: NeilBrown <neil@brown.name>