make: be careful of pane disappearing when doc deleted.
I don't really want to delete the doc, I want to re-used it
so current panes continue to show.  But undo means the doc
can get very big.

So for now, continue to destroy it and make sure we don't
end up referencing a pane after the destruction has happened.

Signed-off-by: NeilBrown <neil@brown.name>