textfill - add simple "fill-mode" command
"M-x fill-mode" will now enable fill-mode in a pane.
Typing space/tab/newline past current width triggers a fill
of the current line.
Width can be set by filling a para with "Num-arg M-q".

There are some weaknesses - sometimes it isn't the last space
that should be replaced.  But it is a start that I can
experiment with.

Signed-off-by: NeilBrown <neil@brown.name>