Delay freeing of pane->handle
Next patch will need to call on a pane very late, to get it
to free ->data.  So we need the handle to persist to the very end.

This means we need a new DAMAGED_DEAD to be set after 'Close' to ensure
nothing else gets called on the pane in the interim.

Signed-off-by: NeilBrown <neil@brown.name>