Delay freeing of pane->data
We already delay the freeing of a pane until it cannot be in use.
But we also sometimes need the ->data to hang around a bit beyond
close too.
So introduce a new "Free" callback to tell a pane to free its
data, and call that just before the pane is freed.

Signed-off-by: NeilBrown <neil@brown.name>