Make: reuse old make-output document
If a make output document exists, clear it instead of deleting
it.  This improves stability of the display.
Also if it is already displayed in a tile, use that one rather
that insisting on an 'other' tile.
We tell the old over-lay pane to remove itself and create a new
clean one for the new output.

Signed-off-by: NeilBrown <neil@brown.name>