Use home_pane_notify() for indirect notifications.
We allow indirect notifications to be send via the doc,
via the window, or globally via the editor.
For 'doc' we use home_pane_notify() to preserve the original
focus.
For window and editor we currently lose the focus.
Change window and editor to use home_pane_notify() just
like 'doc'.

Signed-off-by: NeilBrown <neil@brown.name>