make: use notifier chain to link output docs together.
Instead of recording the name of the current make/grep output doc
in a global attr, link all such docs together in a notifier
change.
Then it is simple to send a notification to ask for the
'next' match.  Each doc can check if a previous doc has
already responded.

Signed-off-by: NeilBrown <neil@brown.name>