display-pygtk: convert from pygtk to pygobject
Apparently pygobject and GTK3 are the way of the future.
So it is time to switch over.

This all seems to work, but there is probably room for
improvement.

In particular, cairo context handling might not be ideal,
and I don't like that I don't get a call back when someone
wants mt clipboard contents.

Signed-off-by: NeilBrown <neil@brown.name>