Review how pane-clear interacts with pixmap
We now request 'pane-clear' to be called on any pane
that will be drawn on.
For pygtk, that allocates a pixmap for the pane.
Drawing in child panes will fall back to the pixmap
of the parent.

Handling in lib-view is not ideal.  It currently results in
two panes for the same depth, which I would rather avoid.
Maybe I should make a separate child pane for each border,
but I'm not sure that is worth it.
Alternately I need to tell the main child reuse the view pixmap.
Still thinking about this...

Signed-off-by: NeilBrown <neil@brown.name>