search: use correct x,y for moving search box.
We currently move the search box if it might obscure
the current match.
However this code uses the focus-relative x,y values
as though they were home-relative.
This currently always works, but that might change.  So
perform the required mapping.
(and optimize and easy case of the mapping while we are there).

Signed-off-by: NeilBrown <neil@brown.name>