render-lines: pass home and focus to all drawing functions.
Drawing will soon use separate sub-panes to restrict the area
that is cleared and re-drawn.
This means that the pane passed as the focus (which the Display
pane will receive) won't be the leaf of the active tree (which is
where we need to send messages).

So use call_pane() (and variants), and pass 'p' and 'focus' consistently.

Signed-off-by: NeilBrown <neil@brown.name>