render-lines: improve some naming.
When calling render_line() there are two specific points in the line
that might be interesting.

There is a location in the marked-up text that is the "cursor", and we
want to know where on the display it is, how big it is, and might want
to draw a box there.

There may also be an x,y location in the display, and we want to know
where in the marked-up text this is, and what attributes are in effect
there.  This often corresponds to the location of a mouse click, and we
currently call it the 'cursor' too.  This is confusing.

So change the name of the xylocation to xypos or similar.  Hopefully
that will reduce my confusion.

Signed-off-by: NeilBrown <neil@brown.name>