render-complete: fix completion of doc names.
This code seemed to assume that doc name are sorted, but they aren't.  So it broke.
Particularly if the doc names were, for example

 doc-text.c
 core-doc.c
 doc-dir.c
 doc-doc.s

in reverse order (because they are examined from last to first),
then when it saw doc-text.c it would set at_start to 1,
then core-doc.c would cause at_start to be set to 2.
This caused all later doc* to be ignored for the purpose of
calculating "common".

Signed-off-by: NeilBrown <neil@brown.name>