render-lines: don't damage self unnecessarily.
When render-lines notices that the point has moved, it
uses "view:changed" to let everyone know.  It even
lets itself know so that highlighted text gets updated.

However this damaged the pane, which isn't really needed in this case.
It causes the pane to be drawn twice, which is a waste.
So add a hack to avoid the unnecessary damage.  Hopefully this
code will be revised soonish as it isn't really render-lines
role to send this notification.

Signed-off-by: NeilBrown <neil@brown.name>