Adjust for python3 and None < int comparisons.
Python2 allows None to be compared (<, > etc) with an int.
Python3 doesn't.

So change the standard 'call' to return 0 for 0, rather than None,
and adjust the few places that depended on a None return.

ret='char' previously returns None both for EOF and 'no response'.
Now it gives 'False' for 'no response'

Signed-off-by: NeilBrown <neil@brown.name>