c-mode: improve indenting of preprocessor directives.
preprocessor directives interrupt the C code but don't affect
it's parsing.  So we need a separate stack for the life of the directive.

looks a lot like a prefix which affects the indent of the following line
in a way we don't want - so explicitly detect and skip it.

Set te base continuation indent to 8.

Signed-off-by: NeilBrown <neil@brown.name>