c-mode: avoid adding extra indent in a couple of cases.
The indent parser can report that an extra indent is permitted.
We only one to add that indent when :Tab is typed at the
start of the text which is currently indented correctly.

If it was a Reindent, not a Tab, then we don't want it.
If cursor was not at the end of the indent, we just want to
move there - the next tab can add the indent.

Signed-off-by: NeilBrown <neil@brown.name>