Allow mode-setting to be done from python.
Mode:set-mode currently requires a static string, and python
cannot provide that.
So change it to dup and free the string.
Also provide a Mode:set-all to make it easier to set numbers
at the same time.
A prefix like :C-x can now be handled in python with something like

def handle_cx(self, key, focus, str, num, num2, **a):
	"handle:K:C-x"
	return focus.call("Mode:set-all", ":CX", num, num2)

Signed-off-by: NeilBrown <neil@brown.name>