c-mode: don't clear comma_ends except via 'pop'.
Once we have entered a comma_ends region, we stay in that mode in
sub-regions, and it is all probably still a list of
values.

e.g. { a, b, {1, 2} }

The , between 1 and 2 should end a 'statement' in just the same way as
the comma between a and b.

Once the state that introduced comma_ends is poped, it will effectively
be cleared.

Signed-off-by: NeilBrown <neil@brown.name>