c-mode: handle macros in value declarators better
This is a bit of a hack specifically to support
the way the python-C-api requires values to be declared.

A variable declarator might contain a macro call which might
contain a trailing comma, which hides the comma from a shallow
parser, so we don't know that it ends a "statement".
We can instead notice the '.' at the start of the next 'statement'
and assume that really is the start of a statement.

Also don't recognize a 'prefix' in the context where a comma ends a
statement.  The concept of 'statement prefix' doesn't apply in that
context.

Signed-off-by: NeilBrown <neil@brown.name>