make/shell: add PWD to environment
Some programs use env[PWD] as the name for '.' if it is correct.  This
might be different from /bin/pwd, and might be preferred.  In particular
some build code includes a full path in results and that can be
different when PWD is different.

So to ensure 'make' gives consistent results when run from a shell or
from edlib, set PWD the same way that the shell does.

Signed-off-by: NeilBrown <neil@brown.name>