make/shell: decode utf-8 one line at a time.
We currently read bytes from the pipe connected to the subprocess in
chunks of up to 1024 bytes, and decode these via utf-8.  It is quite
possible that such a block ends with part of a multi-byte utf-8 code
pointer, so decoding can fail.

Change to instead break the incoming stream at each newline (which
cannot be in the middle of a utf-8 codepoint) and decode the
sets-of-lines.

Signed-off-by: NeilBrown <neil@brown.name>