text: fix capure of doc:revisit notification.
a text doc wants to respond to doc:revisit notification, but never
registers for them - as it is odd to register for a notification that
you yourself send out.

Rather it capture when something asks it to send out such a
notification, but it currently doesn't capture that request properly.

Change the key-add to capture the request, not the notification.

Signed-off-by: NeilBrown <neil@brown.name>