input: invert rule for reporting Release.
Previously we didn't report the Release if a Click event was handled.
Change that to *only* report a Release if a Press event is handled.

This means that if neither click or press are handled, no Release is sent.
This is useful if Press is ignored when some modifiers are present.

Signed-off-by: NeilBrown <neil@brown.name>