ncurses: add modifiers to Click, and handle in input.
1/ ncurses now adds :M:C:S modifiers to Press (not Release)
2/ pygtk now skips the modifiers for Release
3/ lib-input now passes those modifiers to :Click and :DPress etc.
4/ mode-emacs access :C:Click-1 for paste

Terminal emulators often capture :S:Press, and window systems might
capture :M:Press, so it is mainly :C:Press that this is useful for
in ncurses.
xfce-terminal doesn't pass Press-2 down even when it should, so I cannot
use that for paste - so add :C:Click-1 as an alternate mouse paste.

Signed-off-by: NeilBrown <neil@brown.name>