lib-input: add selection-management interface.
This selection-management interface allows a primary
selection to be tracked between panes.

This is expected to be used with the copybuf functionality as the
selection may get converted into a copy.

There needs to be a separate selection for each display.  Currently the
code is in lib-input which is per-window.  This is probably adequate,
but some day I should make it per-display.

Also change mode-emacs to use this interface to work
with mouse selections.

Signed-off-by: NeilBrown <neil@brown.name>