python: remove handler/refer arg to Pane.__init__
The handler that things are refered to is no longer
used, so remove it.

This means that edlib.Pane and edlib.Doc and *only* be used as a base
class.  Creating a stand-alone instance is no longer useful as it
cannot accept commands.

If this ever becomes a problem, it won't be hard to revert this commit,
or do something else appropriate in the context.

Signed-off-by: NeilBrown <neil@brown.name>