viewer: send any unknown characters to doc as commands.
Part of the point of 'viewer' mode is that all normal characters
can become commands - they no longer insert.
It can be useful to send these to the document.

So translate such keys to doc:cmd-key or doc:cmd:key and
resend.

Change doc:docs to request the viewer, and to accept commands rather
than doc:replace.

If a doc:cmd returns 2 or 3, treat that as a request to move up or down.
This seems like a special case for doc-docs, but it will work for
doc-dir as well, and might be more generally useful.

Add 'y' and 'n' docs command to say 'yes' or 'no' to an implicit 'save
this file?'

Signed-off-by: NeilBrown <neil@brown.name>