emacs: use attr on point to check if movement during double-click.
If the mouse moves during a double-click we want to treat the second
click as a normal click - not doubled.  i.e. don't select words.

We current check if point is same as click position, but that assumes
that point hasn't moved.  However to highlight words asap, we need to
move point to the end of a word (future patch).

So instead, record x,y as an attribute on point and use that to check
for change.

Signed-off-by: NeilBrown <neil@brown.name>