emacs: track mouse movement when selecting.
Tracking a changing selection requires an extra mark to remember where
the selection started.  This is different from the start of the
selection as the later might be the start of end of a word or line, and
as the mouse moves it might need to change from the start to the end,
and back.

So use one of the newly added extra marks and record on it whether we
are selecting char, word, or line, and whether movement should be
tracked (as button is down).

Signed-off-by: NeilBrown <neil@brown.name>