ncurses: fix double-click detection.
It seems that we need a small mouseinterval > 0 in order for ncurses not
to get confused about mouse clicks.
With the previous code, A triple-tap-drag on my touchpad, which comes
through X11 as press/release press/release press - motion - release
gets reported by ncurses as press/press/release motion release.

Weird, but it works OK with the new code.

This allows us to remove some ncurses work-arounds from lib-input.

Signed-off-by: NeilBrown <neil@brown.name>