Change point:moved to point:moving, and make it more reliable.
Sometimes we want to know a point movement because we want to see the
before and after.  We can always get the 'after' when we need it,
so it would be nice to be able to capture the 'before' when movement
starts.

So change to point:moving, and generate it before any change.

Also make sure it is generated when the point is completely valid, so
mark_dup() can do something useful.  Previously mark_dup() wasn't safe.

So review all doc:step and doc:set-ref to ensure some appropriate
interface is called first.

Also, default all points to appear to have repored already, so anyone
interested in notifications must call mark_ack() first.

Signed-off-by: NeilBrown <neil@brown.name>