render-lines: fix target-column tracking.
Were we tracking the target column by seeing of point had changed.
But target column can change without moving point (with-respect-to a
saved copy) if we insert before both point and the saved copy.

So we need to reset target column on changes too.

While we are are it, use the new point:moving notification rather than
tracking the change outselves.

Signed-off-by: NeilBrown <neil@brown.name>