emacs: beginnings of command-completion for M-x
We add a command to *every* keymap which will return
all the keys with given prefix that are in keymaps
for the home pane.

This can give duplicates when a pane has a chain of keymaps, but I think
we need to live with that.
It doesn't work with mode-emacs commands, which will be fixed
eventually.

Currently we only support completion of a unique prefix.  Popping up a
list of options will come later.

Signed-off-by: NeilBrown <neil@brown.name>