render-lines: fix possible use-after-free
As call_render_lines() can free other vmarks (other than the one passed
in), we need to be careful of any other vmarks we hold at the same time.
Mostly we don't hold one, but when we do, pass a pointer to it
so call_render_lines() can provide an alternate vmark if it wants to
free this one.

Signed-off-by: NeilBrown <neil@brown.name>