render-lines: ensure text inserted at top appears.
If text is inserted just before the visible section, move the visible
section to the start of that text so that it all appears.
This might be a bit simplistic, but it seems to work.
If 'point' is near the bottom it still might not appear as keeping point
visible is a priority - but I don't think that will be a problem.

Signed-off-by: NeilBrown <neil@brown.name>