complete: don't treat empty prefix like no prefix
We current handle zero-length prefix by setting common_pre to NULL,
which is then treated as "we haven't seen any prefix, so take whatever
we find".
Clearly this is wrong.  Once we see a match for which there is no
prefix, we must presist in having a zero length prefix.

So remove unnecessary condition.

Signed-off-by: NeilBrown <neil@brown.name>