make: avoid reparsing lines
Currently a parse always starts from the last match.
But, particularly with 'make' output, there could be lots
of lines that don't match.
We currently parse them multiple times.  This is particularly
relevant for 'Entering directory' lines.  We only want to parse
them once.

So keep track of the pos of the last parse, and restart from there.

Signed-off-by: NeilBrown <neil@brown.name>