make: allow separate make output for separate directories.
'make' will only re-use a make doc if it is for a compatiable directory.
Thus there can be several active at a time in different projects.
When you "make" in a project, the relevant doc is used.
If "make" is still running there, we display the output but don't
kill it or start a new make.  It can be killed with :C-g and then
restarted easily.

Signed-off-by: NeilBrown <neil@brown.name>