doc:step shouldn't step over marks unnecessarily.
Currently doc:step always steps passed any marks with the same ref.
This shouldn't be needed and isn't implemented consistently (some docs
use mark_to_mark() to leave it in an undefined position in all the
marks).
I'ld rathe caller explicitly used mark_step() when that is needed, and
not depend on any default.  So remove the old default and only
step over marks that need to be stepped over.

I particularly want this for future patch were we can share
responsibilty for part of the ref between panes.

Signed-off-by: NeilBrown <neil@brown.name>