linefilter: send message when view becomes empty.
When the filtered view of the document becomes empty, send a message, as
some users might want to react to that event.

Also doc:render-line to mark will now move the mark forward if it turned
out to be in dead space before the line.  This rendering keeps the point
in visible part of the doc.

Signed-off-by: NeilBrown <neil@brown.name>