pane: don't continue child loop after calling a pane.
Any time we make a call on a pane, artbitrary things might happen,
including panes being closed or moved.
So if we do that in a loop, it isn't safe to continue the loop,
even when using a _safe varient.

So add a DAMAGED_ flag to say "Haven't done this one yet",
set it on everything before the loop, then clear it as things are
processed, restarting the loop as needed.

Signed-off-by: NeilBrown <neil@brown.name>