search: move cursor to end on C-w etc
Whe C-w or C-c is typed the new chars are added at the end of the search
string, but the cursor currently stays were it is, which is confusing.
I don't think there is really ever a need for it to start where it is,
so always move to the end first.

Signed-off-by: NeilBrown <neil@brown.name>