autosave: move follow-symlink functionality to doc-dir
Doc-dir can now be asked to open a file by first performing
a readlink and opening that - so that the name edlib knows
the doc by is the linked name.
Use that functionality in autosave instead of including it
there, and also handle the normal directory open command (f, o, Enter)
for autosave dir.

Signed-off-by: NeilBrown <neil@brown.name>