format: add render-format2
render-format2 is an alternate rendering which uses a new
feature of sharing control over mark-refs between a render pane and the doc.
The render pane can use the 'i' field to track position in a line which
is based on the 'p' field provided by the doc.

This allows selection of text and search to work seemlessly.
I'm not really sure if I want this, as it might be better for search
and select to be integrated differently with more knowledge of the
underlying doc.  We'll see.
For now I can experiment with both and see what I think.

Signed-off-by: NeilBrown <neil@brown.name>