pane: handle post-order damage better.
Both 'review' and 'refresh' handling (and probably resize)
can trigger postorder damaged.  Currently we check for it
twice, but that results in double-refreshes.

Change to treat POSTORDER and a clearly distinct damage type
with POSTORDER_CHILD tracking whether any children need it,
and do all the postorder handling at the end of refresh

Signed-off-by: NeilBrown <neil@brown.name>