Revise pane_resize and DAMAGED_SIZE handling.
DAMAGED_SIZE means that the pane needs to reconsider its own size and/or
that of its children, depending on who is reponsible.

When changing size, pane_resize() should always be used.  It sets
DAMAGED_SIZE if needed so propagating to children will follow a common
path.

DAMAGED_SIZE is now set automatically where children are added, so many
places that currently set it, don't need to any more.

The Notify:resize signal is now sent by pane_resize() (not when
DAMAGED_SIZE it set) and the handler should generally just set
DAMAGED_SIZE on the target.

DAMAGE_CONTENT should not be set when the size changes - that has a
separate use.

Signed-off-by: NeilBrown <neil@brown.name>