Introduce pane_mapxy()
The pane_mapxy() maps co-ordinates from one pane to another.
This is similar to pane_map_xy() - which is being removed - but the
interface is hopefully simpler.
x,y are passed in by-value, and a 'struct xy' is returned.
This seems to simplify some of the calling code.

Also introduce Pane.mapxy for python and use instead of
Pane.abs().  Discard Pane.abs and Pane.rel.
pane_rel() also becomes static.

Signed-off-by: NeilBrown <neil@brown.name>