view: call pane-clear from Refresh, not Refresh-size
The new rules is that drawing and clearings should only
happen from a Refresh handler.

That didn't work before because when view cleared
the pane, it would erase what had been drawn over the
main part of its pane.  But now that we use panels for ncurses, that
isn't a problem any more.

Signed-off-by: NeilBrown <neil@brown.name>