pane: rewrite pane_do_resize()
As a "Refresh:size" handler could potentially close things,
and because it could potentially resize a parent, I want to
always search up from the root to find a thing to resize, then resize
it, then start again.  This is a bit more work, but should be a lot more
robust and handle weird dependencies.

To support this, pane_damaged() now seds DAMAGED_SIZE_CHILD on the pane
that gets DAMAGED_SIZE, not just on its ancestors.
Also, we don't automatically set DAMAGED_CONTENT any more, panes that
do any drawing need to explicitly after of a refresh themselves.

If we try to resize more than 1000 panes, we assume something is wrong
and abort the resizing.

Displays now have a depth of 1 so that there parent never tries to
resize them.

Signed-off-by: NeilBrown <neil@brown.name>