Move point-movement handling from render-lines to core-doc
It never made sense to have this in render-lines.
I'm not entirely sure core-doc is best, but I think it is better.

We use DAMAGED_VIEW to record that something is needed, then
view:changed to ensure the change is visible.

This doesn't quite work yet as emacs doesn't set "selection:active" yet.
This only affect changes to selection that cover multiple lines.

Signed-off-by: NeilBrown <neil@brown.name>