pane: Only call Refresh:view on panes which request it.
Rather than calling Refresh:view on the focus panes were it might be
wanted, call is specifically on just those panes that actually request
it by setting DAMAGED_VIEW.  This is more consistent with the new
meaning and avoids panes getting Refresh:view called when they don't
want it.

Consequently, the handlers don't need to be careful to return 0 any more
as there is no fall-through.

Signed-off-by: NeilBrown <neil@brown.name>