render-lines: set ->cx ->cy more accurately.
A char beyond visible e-o-l can be rendered as two cell,
the "\" at eol, and the actual char at start of next line.
In this case we want the cursor on the last cell.

A tab char can also be rendered as multiple cells, each a space.
On this case we want the cursor on the first cell.

Resolve these by using the first cell on the last line.
Now we can use ->cx ->cy of the focus cell for placing the cursor
in ncurses.

Signed-off-by: NeilBrown <neil@brown.name>