keymap: simplify hashing.
Instead of calculating a hash for every possible prefix length on a key
lookup, calculate just 2 - one to the first [:-], one to the end.
The first is used when a fill key exists in the keymap.
The second when a prefix or range is present.

This increases the number of false positives from 5.5% to 6.8% on my
current test suite, but substantially reduces the complexity and hash
storage.

Signed-off-by: NeilBrown <neil@brown.name>