ncurses: don't assume Draw pane same size are clear pane.
The pane that Draw:text is called on might be different to the one
that was cleared, and so might not have a panel.
Have to find the pane with the panel (already do), and mustn't assume
the geometry of the two is the same - co-ord needs to be mapped.

Signed-off-by: NeilBrown <neil@brown.name>