renderline: don't pass back cursor size.
Instead of passing back cursor size from renderline, just
assume it is the width of an M - a size we already need.
This fixes a bug as the cursor size was sometimes being passed
back as -1.  This is what the 02-grep test is being updated.

As a result, the "render-done" callback is no longer needed.

Signed-off-by: NeilBrown <neil@brown.name>