emacs: case-change: only set num2 when there is a case change.
If a case-change request didn't change anything, then don't record this
as the first of a possible series of cache changes, for the purposes of
undo.

If a change *was* made, or there has already been a recent change, then
do set num2.
This ensure the "first" text-edit is properly marked.

Signed-off-by: NeilBrown <neil@brown.name>