render-line: change 2-line spacing at EOF.
"line_height" isn't really well defined (lines in a page might be of
different heights), so I'm phasing it out.

When moving to EOF, we currently leave 2 blankes lines after EOF to make
EOF stand out more.  This work well for 20-30 line panes, but is a waste
in smaller panes.
Change this to leaving 10% (rounded down) of the pane height.

This requires some tests to be refreshed as they have half-height panes
and move to EOF.

Signed-off-by: NeilBrown <neil@brown.name>