renderlines: rename skip_lines to skip_height
The units of 'skip_lines' is the same as pane->h, so it isn't
text lines and hence the name is confusing.
So change it to 'skip-height'.

A few places were treating it as textlines, which was a bug.
Change those to include a factor of rl->line_height to make it
really work as text lines.

Signed-off-by: NeilBrown <neil@brown.name>