renderlines: change how vline==0 is handled.
Rather than doing a calculation based on line_height,
simply set 'top' and 'bot' to NULL.  A vline of zero
will otherwise be ignored, so above and below will
balance out.

This will change behaviour when cursor is at end of file.
Previously, EOF would have been placed in middle of page.
Now it will remain near the end.
I think this is OK.  It is at least consistent with start-of-file,
which is always at top of page.

Signed-off-by: NeilBrown <neil@brown.name>