render-line: line_height is not a global.
Each line could potentially have a different height, so having a global
line_height for a doc is wrong.
So extract line_height when we measure a line, and use it as
necessary.

The "skip_height" sometimes needs to be treated as lines, so
store a "skip_line_height" together with that.

This removes the last user of the render-line callback, so it
gets removed.

Signed-off-by: NeilBrown <neil@brown.name>