render-lines: only send render:reposition when there is good reason too.
We have a '->repositioned' flag, but we don't use it any more.
We should, to avoid needless work.
So if we know we have repositioned, or if we have been asked to
reposition even if we didn't actually need to, then send
the responsition notification.  Else don't bother.

This removes some noise from a couple of tests.

Signed-off-by: NeilBrown <neil@brown.name>