lib-renderline: don't adjust width on page.
The current width calculation is wrong for TABs, and
would need to include the right margin where a wrap-character
goes.
We don't actually use this width in any important way, but
making the pane too narrow causes confusion.

So keep the helper pane the same width as the main pane.

Signed-off-by: NeilBrown <neil@brown.name>