linefilter: send view:changed if filter info changes.
When filter info isn't set explicitly, we regularly check the
attrivbutes which set it implicitly.
If they change, send view:changed.

This is particularly import if a pane closer to the root depends on the
rendering that this pane does.  If something else triggers view:changed,
the other pane will re-assess before this pane has a chance to, so the
other pane might need to re-assess again.

Signed-off-by: NeilBrown <neil@brown.name>