renderline: centre image based on parent pane.
As we resize an image pane to position the image,
we cannot use the size to decide where to center the
image.
So use p->parent instead.

I'm not sure this is the right final answer, but it helps until
I have a chance to think through what I really want in terms
of the size of these panes.

Signed-off-by: NeilBrown <neil@brown.name>