render-lines: separate out Refresh:view from Refresh.
render-lines currently does lots of work in "Refresh",
possible choosing a different set of lines to draw.

Some of this conceptually fits better in Refresh:view, and choosing
new place to draw will trigger render:reposition, which might
cause other panes to re-assess their view.

So create a separate 'render_lines_revise' which handles Refresh:view.
This assesss if the current top is acceptable - will it display the
point if that is needed - and if so, checks the heights of all subpanes
and makes sure thare are all valid and correctly positioned.

If the cursor is needed but wouldn't be displayed 'find_lines' is called
which chooses a new top-of-pane line.

shift-left changes are handed by the new render_lines_revise, and it
does so while just examining the cursor line.  We no longer redraw after
the pane if a shift is found to be needed.

By the time "Refresh" is called, all sub-panes are known to be correct,
as it that is needed is to drawn the content on the display.

These changes result in redraws sometimes happening in a different
order.
As the way find_lines chooses a new top depends on the previous set of
displayed lines, this will sometimes result in a slightly different
'top' being chosen.
Also, the target column is no preserved during page-down or page-up.
Consequently some of the tests need to be refreshed.

Signed-off-by: NeilBrown <neil@brown.name>