render-lines: switch to clearing each sub-pane.
By clearing each subpane, they each get a separate image
cached in the display.  This will allow us to only redraw
when the line itself changes.

This fixes a bug, so one test needs to be refreshed.
If the last line on the pane wraps and doesn't all fit,
the wrap char '\' was getting lost.  We now don't lose it.

Signed-off-by: NeilBrown <neil@brown.name>