diff/make: use Move-EOL to find a given line.
Move-Line is more heavy-weight as 'render-lines' tries to keep
the column and generally does extra work.  It also causes
problems at the moment.

This fixes a bug where a line jumped to by 'diff' often appears at the
top of the pane.  So the 'diff' test needs to be refreshed.

Signed-off-by: NeilBrown <neil@brown.name>