Introduce DocLeaf and use for diff and make
When diff or make/grep want to jump to a line in a document, and
the document is visible, they currently re-attach it to the same pane,
losing context.
Change to leave the doc where it is.  pane_leaf() on the result of
DocPane doesn't do the right thing if the pane has a popup, so
provide DocLeaf which does the right thing.
Not sure if this is the best design.

This changes the behaviour (for the better) for serveral tests, so they
get refresh.

Signed-off-by: NeilBrown <neil@brown.name>