render-lines: don't move line-pane when measuring it.
We currently move a subpane to 0,0 when measuring it.
This is not really expected from the name "measure_line()"
and causes a problem because we sometimes measure the last
line of a pane to check where the cursor is.

So change to always leave ->x and ->y unchanged.  Other code will
update these values when needed.

Signed-off-by: NeilBrown <neil@brown.name>