render-lines: don't force find-lines() when appending to doc.
When we append to a doc, the point will move after the previously
displayed content.  This currently causes find_lines() to be required,
but that isn't really what we want.  We want to still keep the
top-of-display stable in that case.

We can detect it because m2->mdata will be non-NULL.  That only happens
when it is end-of-file.

Signed-off-by: NeilBrown <neil@brown.name>