render-lines: limit space at EOF for small movements.
It is nice to have extra space at EOF to make it obvious when there is a
largish movement to get there.  But if alredy at EOF, the extra space
doesn't help so much and can be annoying.
So suppress any space added after EOF that is more than the distance
from the old end-of-display to EOF.

Signed-off-by: NeilBrown <neil@brown.name>