shell: add default viewer that follows output.
Sometimes I want shell output to scroll so that I can see the end.
I'm not yet sure that I always want this, and I know I don't want
it for diff, so diff over-rides it.

I'll try it an see what I think.  Certainly having it as an option seems
to make sense.

Signed-off-by: NeilBrown <neil@brown.name>