rexel: change how ignoring-case is implemented.
Rather than have a 'ignore case' setting that is global for the whole
pattern, we now allow it to be applied to ranges within the pattern
using new RXL commands: REC_IGNCASE and REC_USECASE.
When performing a match, we scan for these and create a bitmap so we can
easily check if any particular rex-el should be ignoring case.

Also add some ignore-case tests to the self-test, and fix so that it
builds.

Currently we only ever put a REC_IGNCASE at the start and never use
REC_USECASE, but that will change.

Signed-off-by: NeilBrown <neil@brown.name>