rexel: change parse_re() from recusive to iterative.
Rather than having parse_re() call itself recusively, only call
parse_branch() and do this iteratively.

The recusion used the stack to remember where each branch ended so that
a fork-to-end could be placed once the end address was known.
Instead, chain the temp-forks together and walk that list updating them
all to point to the end.

This will allow my to introduce a syntax that is permitted at the start
of an re, but not at the start of other branches.

Signed-off-by: NeilBrown <neil@brown.name>