rexel: support re mods at start of re.
If a regular-expression (including within "()") starts with '?'
then this provides modifiers for parsing/matching.
There must be a subsequent colon, and all between '?' and ':'
are mods.
A strong of digits at the end is a length (in characters) of a literal
pattern, which can contain any char, and must consume the whole re.
Other modifiers are:
 i - ignore case in this re
 l - lax matching for space and dash
 s - be sensitive to case and don't do lax matching.

Note that these are subject to change after some experience.

Signed-off-by: NeilBrown <neil@brown.name>