rexel: support \B - non-word boundary
Now that non-word boundaries are reported, we can reliable match them.
In most cases, a report for non-word-boundary is uninteresting, but
there are likely to be a lot of them.  So sort REC_NOWBRK at the front
of the linked list, and optimize away if no-one cares about the
non-word-boundary.

Signed-off-by: NeilBrown <neil@brown.name>