search: remove incorrect abort test.
We currently abort the search if we have previously found a match, but
the current position doesn't match.
This is incorrect as (abc)+ matching against abcabc will abort
when it sees the second a doesn't result in a match.
We should always continue until a -2 return meaning that there are no
more possible matches.

Signed-off-by: NeilBrown <neil@brown.name>