rexel: ensure backtracking finds longest first
Forward forks are now treated differently to backward forks.
A forward fork we follow later, a backward fork we follow first.
This means we will find the longest match for any repeating
pattern first.

Also abort as soon as a match is found, so that the record
is correct for a match.

Signed-off-by: NeilBrown <neil@brown.name>