rexel: make clear difference between priority and delayed forks.
For backtrack-matching it makes a difference whether we follow a fork
the first time it is seen, or the second.
We were following the first time for backward jumps and the second
for forward jumps and that serves most purposes.
But to implement '?*' which is a non-greedy match we need to
follow a backward jump second.
So halve the supported size of jumps and use a bit to record whether
we want an early jump or a late one.

Signed-off-by: NeilBrown <neil@brown.name>